[DEV-6377] Period picker implementation:
1. WithLatestFy now provides the actual array from the api
of latest submission periods
2. capturing & validating period selection in url
3. Providing enabled/disabled styles
4. Using data mapping to get the periods in shape

NOTE on (3) above:
FY's b4 2020 only allow for quarter selection.
Special styles apply for this case per the mock.
To conform to these styles, we have to style
sibling elements. Sibling elements can only
be styled when they come after the element you
are selecting, so we put all the last periods of
the quarter at the top of the list, then use
the css property 'order' to make the order correct.

This is kind of hacky and warrants this elaborate
note.